refactor(040): migra packages/core/src/chatbot para TypeScript
