+++ packages/core/src/chatbot/chatbotUiText.ts
